backport contrib test file restructuring
